Move WorkSpaceLayout to Layout folder
