+++ Source/SelfService/Web/components/layout/workSpaceLayoutLinks.tsx
@@ -5,11 +5,11 @@ import React from 'react';
 
 import { Link, useLocation } from 'react-router-dom';
 
-import { useGlobalContext } from '../../../context/globalContext';
+import { useGlobalContext } from '../../context/globalContext';
 
 import { getPrimaryNavigationItems, getSecondaryNavigationItems, getSidePanelItems, LayoutProps, MenuListProps, MenuItemProps } from '@dolittle/design-system';
 
-import { ApplicationChanger } from '../../../spaces/applications/applicationChanger';
+import { ApplicationChanger } from '../../spaces/applications/applicationChanger';
 
 const PrimaryNavigation = () => {
     const location = useLocation();
